test(backend): lock GitHub Contents base64 decode without Buffer
Prevents skill sync from regressing to Node Buffer in Convex isolate.
