+++ packages/backend/tests/decodeGitHubContent.test.ts
@@ -1,0 +1,16 @@
+import { expect, test } from "vitest";
+import { decodeGitHubContent } from "../convex/_repoSkills/decodeGitHubContent";
+
+test("decodeGitHubContent decodes base64 without Node Buffer", () => {
+  // GitHub Contents API returns base64; Convex isolate has no Buffer.
+  const text = "hello skills\n";
+  const encoded = btoa(text);
+  expect(decodeGitHubContent(encoded)).toBe(text);
+});
+
+test("decodeGitHubContent strips newlines from wrapped base64", () => {
+  const text = "line1\nline2";
+  const encoded = btoa(text);
+  const wrapped = `${encoded.slice(0, 8)}\n${encoded.slice(8)}`;
+  expect(decodeGitHubContent(wrapped)).toBe(text);
+});
